Moving acknowledgeJob call to checkout method
When multiple CodePipeline actions are using same project, and
a new change is detected at the same time, then some of the builds
are not invoked after job was acknowledged. We are moving acknowledgeJob
call right before triggerring the build so that, if checkout method
is not called, then job won't be acknowledged and can be retried after 30sec.