Adding proxy settings in the clientConfig while initializing S3 client
The bug was introduced in 0.16 release where we stopped initializing s3 client
with proxy settings. Due to that, customers who use plugin in the private network
can't access S3 buckets. This fix add proxy settings, and also remove AWSS3S4SignerType
as it's going to be defauly SigV4 in SDK 1.11.x