Add support to input artifact names for locations in publisher
Originally, the publisher upload outputs to artifact locations based on the order returned by pollForJobs,
which is not predictable.

This commit is adding artifact name to output location to allow users to enter the name of the artifact location
they want to upload to. The new field "Artifact Name" is optional. Users could choose not to input any artifact
name, which will default to the previous behavior, or enter artifact name for each location.